requirements.txt
baidu-aip==2.2.12.0
certifi==2018.11.29
chardet==3.0.4
idna==2.8
PyExecJS==1.5.1
PyMySQL==0.9.3
pyttsx3==2.7
pywin32==224
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wincertstore==0.2
